+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Correctly add HandleSensorCapability
+
 ## 1.0.46 (2022-08-13)
 
 * (theimo1221) Expose shutter position control in API